refactor: move ProjectName to models, remove from metadata interface
