+++ common/src/lib.rs
@@ -16,8 +16,6 @@ pub mod log;
 pub use log::LogItem;
 #[cfg(feature = "models")]
 pub mod models;
-#[cfg(feature = "service")]
-pub mod project;
 pub mod resource;
 pub mod secrets;
 pub use secrets::{Secret, SecretStore};
