+++ gateway/src/acme.rs
@@ -14,8 +14,7 @@ use instant_acme::{
     Identifier, KeyAuthorization, LetsEncrypt, NewAccount, NewOrder, Order, OrderStatus,
 };
 use rcgen::{Certificate, CertificateParams, DistinguishedName};
-use shuttle_common::project::ProjectName;
-
+use shuttle_common::models::project::ProjectName;
 use tokio::sync::Mutex;
 use tokio::time::sleep;
 use tower::{Layer, Service};
