+++ gateway/src/auth.rs
@@ -5,7 +5,7 @@ use axum::extract::{FromRef, FromRequestParts, Path};
 use axum::http::request::Parts;
 use serde::{Deserialize, Serialize};
 use shuttle_common::claims::{Claim, Scope};
-use shuttle_common::project::ProjectName;
+use shuttle_common::models::project::ProjectName;
 use tracing::{trace, Span};
 
 use crate::api::latest::RouterState;
