+++ runtime/src/provisioner_factory.rs
@@ -9,12 +9,12 @@ use shuttle_common::{
     DatabaseReadyInfo,
 };
 use shuttle_proto::provisioner::{provisioner_client::ProvisionerClient, DatabaseRequest};
-use shuttle_service::{DeploymentMetadata, Environment, Factory, ProjectName};
+use shuttle_service::{DeploymentMetadata, Environment, Factory};
 use tonic::{transport::Channel, Request};
 
 /// A factory (service locator) which goes through the provisioner crate
 pub struct ProvisionerFactory {
-    service_name: ProjectName,
+    service_name: String,
     provisioner_client: ProvisionerClient<ClaimService<InjectPropagation<Channel>>>,
     secrets: BTreeMap<String, Secret<String>>,
     env: Environment,
@@ -24,7 +24,7 @@ pub struct ProvisionerFactory {
 impl ProvisionerFactory {
     pub(crate) fn new(
         provisioner_client: ProvisionerClient<ClaimService<InjectPropagation<Channel>>>,
-        service_name: ProjectName,
+        service_name: String,
         secrets: BTreeMap<String, Secret<String>>,
         env: Environment,
         claim: Option<Claim>,
@@ -75,7 +75,7 @@ impl Factory for ProvisionerFactory {
     fn get_metadata(&self) -> DeploymentMetadata {
         DeploymentMetadata {
             env: self.env,
-            project_name: self.service_name.clone(),
+            project_name: self.service_name.to_string(),
             service_name: self.service_name.to_string(),
             storage_path: PathBuf::from(STORAGE_DIRNAME),
         }
